Day 2 Changes
Added to the calcOrder function to calculate the order subtotal, the cost of the total order, and to store the order details. Created the formatNumber function to format a numeric value using the local numeric format to the number of decimal places specified by decimals. Also created the formatUSACurrency function which formats it's parameter to U.S.A currency.