Day 4 Changes
Created the runSubmit() function that runs validation tests when the submit button is clicked. Created each of the validation test functions that validates that the user has specified the name on the credit card, validates that the user has selected a credit card type, validates that the user has entered a valid and legitimate card number, validates that the user has selected the expiration month of the credit card, validates that the user has selected the expiration year of the credit card, and validates the credit card CVC number.